backend/requirements.txt
fastapi==0.135.1
google-generativeai==0.8.6
pytest==9.0.2
pytest-mock==3.15.1
python-dotenv==1.2.2
uvicorn==0.42.0
